XPathSyntaxMap: Only use normalize-space when needed
If the field is not a string, we don't want to append
"/normalize-space(text())", because then a comparison will not work as
we want. For example; if '/a/b' is a boolean,
'/a/b/normalize-space(text()) = true()' causes an error "Cannot compare
xs:string to xs:boolean".

Update unit tests to match this change.