xpath: Fix handling of missing duration field (TEDEFO-965)
If the duration field is not present, the result of the concat is 'P',
which causes an error with some XPath engines.

Instead, when the field is not present, the XPath now just return the
empty sequence (). As comparisons with the empty sequence is always
false, this gives us the expect result for the expression.

Adapt affected unit tests.