Publish via the new Maven Central Portal
Also use the options of the setup-java action to define the credentials
for the Maven Central Portal, instead of having them in a settings.xml
file.
Move definition of the repository to use for snapshot dependencies to
pom.xml. Having it in settings.xml is not useful, and reuired action by
every developer.
Update README accordingly.